HomeWork Update 1

+++ Completed Projects/ft_printf/ft_printf.h
@@ -21,7 +21,7 @@ int		ft_printf(const char *format, ...);
 void	racc_format_check(va_list arg, char c, int *counter);
 void	racc_putchar(char c, int *counter);
 void	racc_putstr(char *s, int *counter);
-void	racc_putnbs(unsigned int n, int *counter, char *base);
+void	racc_putnbs(unsigned long long n, int *counter, char *base);
 void	racc_putadrs(uintptr_t n, int *counter, char *base);
 
 #endif
